// src/types/api.ts
export interface ModelService {
    cover: number
    created_at: string
    duration: number
    enable_show_in_service: boolean
    enable_trigger_task: boolean
    /** 失败时执行的触发任务id */
    fail_trigger_tasks: number[]
    id: number
    latency_notify: boolean
    max_latency: number
    min_latency: number
    name: string
    /** 当前服务监控所属的通知组 ID */
    notification_group_id: number
    notify: boolean
    /** 恢复时执行的触发任务id */
    recover_trigger_tasks: number[]
    skip_servers: Record<string, boolean>
    target: string
    type: number
    updated_at: string
}